Added Projects section to the page
The projects section starts with a loading image, and after a rest call to github returns with my account's repositories, the image is collapsed and tiles with each repository's informations are loaded. Octicons was added for the nice GitHub icons. All the calls are made with vanilla JS. As we just use a few functions, it would make no sense to use jquery. All 'Javascript' code is made as coffeescript. Gears images for the loading were added.